Move UI functionality to SDK
