+++ packages/snaps-controllers/coverage.json
@@ -1,6 +1,6 @@
 {
   "branches": 90.26,
   "functions": 96.32,
-  "lines": 97.22,
+  "lines": 97.21,
   "statements": 96.9
 }
