+++ packages/snaps-sdk/jest.config.js
@@ -3,6 +3,8 @@ const deepmerge = require('deepmerge');
 const baseConfig = require('../../jest.config.base');
 
 module.exports = deepmerge(baseConfig, {
+  collectCoverageFrom: ['!./src/**/index.ts'],
+
   coverageThreshold: {
     global: {
       branches: 100,
