+++ packages/snaps-sdk/src/index.ts
@@ -18,3 +18,4 @@ export type {
 
 export * from './errors';
 export * from './types';
+export * from './ui';
