+++ packages/snaps-sdk/src/internals/errors.ts
@@ -11,6 +11,7 @@ export const SNAP_ERROR_MESSAGE = 'Snap Error';
  *
  * @param error - The error to get the message from.
  * @returns The error message.
+ * @internal
  */
 export function getErrorMessage(error: unknown) {
   if (
@@ -30,6 +31,7 @@ export function getErrorMessage(error: unknown) {
  * @param error - The error to get the stack from.
  * @returns The error stack, or undefined if the error does not have a valid
  * stack.
+ * @internal
  */
 export function getErrorStack(error: unknown) {
   if (
@@ -48,6 +50,7 @@ export function getErrorStack(error: unknown) {
  *
  * @param error - The error to get the code from.
  * @returns The error code, or `-32603` if the error does not have a valid code.
+ * @internal
  */
 export function getErrorCode(error: unknown) {
   if (
@@ -68,6 +71,7 @@ export function getErrorCode(error: unknown) {
  * @param error - The error to get the data from.
  * @returns The error data, or an empty object if the error does not have valid
  * data.
+ * @internal
  */
 export function getErrorData(error: unknown) {
   if (
