+++ packages/snaps-sdk/src/ui/components/heading.test.ts
@@ -1,0 +1,40 @@
+import { NodeType } from '../nodes';
+import { heading } from './heading';
+
+describe('heading', () => {
+  it('creates a heading component', () => {
+    expect(heading({ value: 'Hello, world!' })).toStrictEqual({
+      type: NodeType.Heading,
+      value: 'Hello, world!',
+    });
+
+    expect(heading({ value: 'foo bar' })).toStrictEqual({
+      type: NodeType.Heading,
+      value: 'foo bar',
+    });
+  });
+
+  it('creates a heading component using the shorthand form', () => {
+    expect(heading('Hello, world!')).toStrictEqual({
+      type: NodeType.Heading,
+      value: 'Hello, world!',
+    });
+
+    expect(heading('foo bar')).toStrictEqual({
+      type: NodeType.Heading,
+      value: 'foo bar',
+    });
+  });
+
+  it('validates the args', () => {
+    // @ts-expect-error - Invalid args.
+    expect(() => heading({ value: 'foo', bar: 'baz' })).toThrow(
+      'Invalid heading component: At path: bar -- Expected a value of type `never`, but received: `"baz"`.',
+    );
+
+    // @ts-expect-error - Invalid args.
+    expect(() => heading({})).toThrow(
+      'Invalid heading component: At path: value -- Expected a string, but received: undefined.',
+    );
+  });
+});
