+++ packages/snaps-sdk/src/ui/components/index.ts
@@ -1,0 +1,9 @@
+export * from './copyable';
+export * from './divider';
+export * from './heading';
+export type { Image } from './image';
+export { image, ImageStruct } from './image';
+export type { Component, Panel } from './panel';
+export { ComponentStruct, panel, PanelStruct } from './panel';
+export * from './spinner';
+export * from './text';
