+++ packages/snaps-sdk/src/ui/components/spinner.test.ts
@@ -1,0 +1,17 @@
+import { NodeType } from '../nodes';
+import { spinner } from './spinner';
+
+describe('spinner', () => {
+  it('creates a spinner component', () => {
+    expect(spinner()).toStrictEqual({
+      type: NodeType.Spinner,
+    });
+  });
+
+  it('validates the args', () => {
+    // @ts-expect-error - Invalid args.
+    expect(() => spinner({ bar: 'baz' })).toThrow(
+      'Invalid spinner component: At path: bar -- Expected a value of type `never`, but received: `"baz"`.',
+    );
+  });
+});
