+++ packages/snaps-sdk/src/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './components';
+export * from './component';
+export { NodeType } from './nodes';
