+++ packages/snaps-ui/src/index.ts
@@ -1,3 +1,2 @@
-export * from './builder';
-export * from './nodes';
-export * from './validation';
+// TODO: Remove this package.
+// eslint-disable-next-line import/unambiguous
