+++ packages/snaps-utils/coverage.json
@@ -1,6 +1,6 @@
 {
-  "branches": 95.83,
-  "functions": 98.99,
-  "lines": 98.65,
-  "statements": 95.49
+  "branches": 95.95,
+  "functions": 99.01,
+  "lines": 98.68,
+  "statements": 95.58
 }
