+++ packages/snaps-utils/src/index.browser.ts
@@ -22,6 +22,7 @@ export * from './snaps';
 export * from './strings';
 export * from './structs';
 export * from './types';
+export * from './ui';
 export * from './validation';
 export * from './versions';
 export * from './virtual-file/index.browser';
