+++ packages/snaps-utils/src/index.ts
@@ -27,6 +27,7 @@ export * from './snaps';
 export * from './strings';
 export * from './structs';
 export * from './types';
+export * from './ui';
 export * from './validation';
 export * from './versions';
 export * from './virtual-file';
